+++ .editorconfig
@@ -1,0 +1,10 @@
+root = true
+
+# defaults
+[*]
+charset = utf-8
+end_of_line = lf
+indent_size = 4
+indent_style = tab
+insert_final_newline = true
+trim_trailing_whitespace = true
