+++ .prettierrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   $schema: 'http://json.schemastore.org/prettierrc',
   trailingComma: 'es5',
-  tabWidth: 2,
+  tabWidth: 4,
   semi: false,
   singleQuote: true,
   arrowParens: 'always',
