+++ src/checks/0ctor.ts
@@ -1,0 +1,3 @@
+export const checkConstructor = (entity: any) => {
+    // TODO:
+}
