+++ src/checks/1compWarnings.ts
@@ -1,0 +1,36 @@
+import { ErrorType, SeverityLevel, ErrorObj } from "../types";
+import { setupErrors } from "../errors";
+
+
+export const checkForCompilationWarnings = (buildInfo: any, severityLevel: SeverityLevel): ErrorObj => {
+    const errors: any = {};
+
+    for (const bi of buildInfo) {
+        const outputErrors = (bi.output as any).errors
+        if (outputErrors && outputErrors.length > 0) {
+            outputErrors.forEach((err: any) => {
+                if (!errors[err.sourceLocation.file]) {
+                    errors[err.sourceLocation.file] = []
+                }
+                const filePath = err.sourceLocation.file
+                const fileComponents = filePath.split('/')
+                const fileName = fileComponents[fileComponents.length - 1]
+
+                errors[err.sourceLocation.file].push({
+                    text: setupErrors(filePath, fileName)(
+                        ErrorType.CompilationWarning,
+                        err.formattedMessage
+                    ),
+                    severityLevel,
+                    type: ErrorType.CompilationWarning,
+                    at: '',
+                    filePath,
+                    fileName,
+                })
+            })
+            break
+        }
+    }
+
+    return errors;
+}
